+++ app/src/main/java/com/a10miaomiao/bilimiao/ui/video/VideoCommentDetailsFragment.kt
@@ -88,7 +88,7 @@ class VideoCommentDetailsFragment : Fragment() {
                         Owner(
                                 face = member.avatar,
                                 name = member.uname,
-                                mid = member.mid.toInt()
+                                mid = member.mid.toLong()
                         )
                 ))
                 MainActivity.of(context)
@@ -151,7 +151,7 @@ class VideoCommentDetailsFragment : Fragment() {
                                 Owner(
                                         face = member.avatar,
                                         name = member.uname,
-                                        mid = member.mid.toInt()
+                                        mid = member.mid.toLong()
                                 )
                         ))
                         MainActivity.of(context)
